requirements.txt
# 版本与 conda 环境 ggd 中 pip 包对齐（pip list --format=freeze，2026-03）。
# 建议在已激活的 ggd 环境中安装：conda activate ggd && pip install -r requirements.txt
# 说明：ggd 环境使用 onnxruntime（CPU/GPU 由该包与驱动决定）；若需强制 GPU 版 ONNX，可自行改换 onnxruntime-gpu 并验证 rapidocr。

# Web
fastapi==0.135.1
uvicorn[standard]==0.42.0
websockets==16.0
pydantic==2.12.5
starlette==0.52.1

# 配置与数据
PyYAML==6.0.3

# A 侧：截屏/模板、音频、FunASR、OCR（Windows）
numpy==2.4.3
opencv-python==4.13.0.92
pywin32==311
PyAudio==0.2.14
torch==2.10.0
torchaudio==2.10.0
funasr==1.3.1
modelscope==1.35.1
rapidocr-onnxruntime==1.4.4
onnxruntime==1.24.4
pillow==12.1.1

# 文档与规则库解析
PyMuPDF==1.27.2.2
openpyxl==3.1.5

# B 侧：阿里云百炼（通义）+ LangGraph + RAG
dashscope==1.25.14
aiosqlite==0.22.1
langgraph==1.1.3
langgraph-checkpoint-sqlite==3.0.3
langgraph-checkpoint==4.0.1
langchain==1.2.13
langchain-core==1.2.20
langchain-community==0.4.1
langchain-chroma==1.1.0
langchain-text-splitters==1.1.1
chromadb==1.5.5
